🧹 Deduplicate GPU + deployment card utilities into lib/gpu.ts
Extract normalizeClusterName (3 copies), hasGPUResourceRequest (2 copies),
and extractVersion → extractImageTag (2 copies) into shared lib/gpu.ts.

Fixes #10131
